address comments for getTextContent
